fix: 서버 컴포넌트에서 getMe 호출 실패 이슈 수정
